Update echo_server.py
I see what you mean about it hanging! I think that this is the line that you would need to change. You're right about closing the `conn` on line 39: the `conn` represents the connection to an individual client and you should close that connection each time after you deal with a client. But then here at the end, when you're ready to close the program, it's the `sock_s` that you want to close: that represents the socket that the server has open to listen for new clients. :)